refactor(test): share CI checkout fixture lifecycle
